A bunch of in progress development that forgot about so I am commiting it and will debug it as I go.

+++ src/Entity/Game.php
@@ -236,6 +236,9 @@ final class Game extends RevisionableContentEntityBase implements GameInterface 
   public function delete() {
     foreach (GameElementType::ELEMENTS as $elementClass) {
       $bundle_id = $elementClass::getBundleByGame($this);
+      if (!$bundle_id) {
+        continue;
+      }
       $bundle = $elementClass::load($bundle_id);
       $bundle->delete();
     }
